remove stream pruner
