+++ src/query/storages/fuse/src/pruning/block_pruner.rs
@@ -48,13 +48,24 @@ impl BlockPruner {
         segment_location: SegmentLocation,
         segment_info: &CompactSegmentInfo,
     ) -> Result<Vec<(BlockMetaIndex, Arc<BlockMeta>)>> {
+        let block_metas = segment_info.block_metas()?;
+        self.stream_pruning(segment_location, block_metas).await
+    }
+
+    // Used for stream pruning.
+    #[async_backtrace::framed]
+    pub async fn stream_pruning(
+        &self,
+        segment_location: SegmentLocation,
+        block_metas: Vec<Arc<BlockMeta>>,
+    ) -> Result<Vec<(BlockMetaIndex, Arc<BlockMeta>)>> {
         if let Some(bloom_pruner) = &self.pruning_ctx.bloom_pruner {
-            self.block_pruning(bloom_pruner, segment_location, segment_info)
+            self.block_pruning(bloom_pruner, segment_location, block_metas)
                 .await
         } else {
             // if no available filter pruners, just prune the blocks by
             // using zone map index, and do not spawn async tasks
-            self.block_pruning_sync(segment_location, segment_info)
+            self.block_pruning_sync(segment_location, block_metas)
         }
     }
 
@@ -64,7 +75,7 @@ impl BlockPruner {
         &self,
         bloom_pruner: &Arc<dyn BloomPruner + Send + Sync>,
         segment_location: SegmentLocation,
-        segment_info: &CompactSegmentInfo,
+        block_metas: Vec<Arc<BlockMeta>>,
     ) -> Result<Vec<(BlockMetaIndex, Arc<BlockMeta>)>> {
         let pruning_stats = self.pruning_ctx.pruning_stats.clone();
         let pruning_runtime = &self.pruning_ctx.pruning_runtime;
@@ -73,11 +84,9 @@ impl BlockPruner {
         let range_pruner = self.pruning_ctx.range_pruner.clone();
         let page_pruner = self.pruning_ctx.page_pruner.clone();
 
-        let segment_block_metas = segment_info.block_metas()?;
-
         let blocks = if let Some(internal_column_pruner) = &self.pruning_ctx.internal_column_pruner
         {
-            segment_block_metas
+            block_metas
                 .iter()
                 .enumerate()
                 .filter(|(_, block)| {
@@ -85,7 +94,7 @@ impl BlockPruner {
                 })
                 .collect::<Vec<_>>()
         } else {
-            segment_block_metas.iter().enumerate().collect()
+            block_metas.iter().enumerate().collect()
         };
 
         let mut blocks = blocks.into_iter();
@@ -188,11 +197,11 @@ impl BlockPruner {
             .map_err(|e| ErrorCode::StorageOther(format!("block pruning failure, {}", e)))?;
 
         let mut result = Vec::with_capacity(joint.len());
-        let block_num = segment_info.summary.block_count as usize;
+        let block_num = block_metas.len();
         for item in joint {
             let (block_idx, keep, range, block_location) = item;
             if keep {
-                let block = segment_block_metas[block_idx].clone();
+                let block = block_metas[block_idx].clone();
 
                 debug_assert_eq!(block_location, block.location.0);
 
@@ -223,7 +232,7 @@ impl BlockPruner {
     fn block_pruning_sync(
         &self,
         segment_location: SegmentLocation,
-        segment_info: &CompactSegmentInfo,
+        block_metas: Vec<Arc<BlockMeta>>,
     ) -> Result<Vec<(BlockMetaIndex, Arc<BlockMeta>)>> {
         let pruning_stats = self.pruning_ctx.pruning_stats.clone();
         let limit_pruner = self.pruning_ctx.limit_pruner.clone();
@@ -232,10 +241,9 @@ impl BlockPruner {
 
         let start = Instant::now();
 
-        let segment_block_metas = segment_info.block_metas()?;
         let blocks = if let Some(internal_column_pruner) = &self.pruning_ctx.internal_column_pruner
         {
-            segment_block_metas
+            block_metas
                 .iter()
                 .enumerate()
                 .filter(|(_, block)| {
@@ -243,10 +251,10 @@ impl BlockPruner {
                 })
                 .collect::<Vec<_>>()
         } else {
-            segment_block_metas.iter().enumerate().collect::<Vec<_>>()
+            block_metas.iter().enumerate().collect::<Vec<_>>()
         };
         let mut result = Vec::with_capacity(blocks.len());
-        let block_num = segment_info.summary.block_count as usize;
+        let block_num = block_metas.len();
         for (block_idx, block_meta) in blocks {
             // Perf.
             {
