+++ src/query/storages/stream/src/lib.rs
@@ -13,8 +13,5 @@
 // limitations under the License.
 
 #![feature(impl_trait_in_assoc_type)]
-
-pub mod stream_pruner;
-pub mod stream_table;
-
 pub mod stream_status_table_func;
+pub mod stream_table;
